feat: 使用最新 dashboard 拓展 API 实现 webhook 路由挂载；清理无意义的防御代码
