+++ plugins/hermes/dashboard/manifest.json
@@ -1,0 +1,6 @@
+{
+  "name": "miloco",
+  "label": "Miloco",
+  "description": "Xiaomi Miloco whole-home AI intelligence",
+  "api": "api.py"
+}
