+++ plugins/hermes/pyproject.toml
@@ -7,11 +7,14 @@ name = "miloco-hermes-plugin"
 version = "2.0.0"
 description = "Xiaomi Miloco Hermes Agent plugin"
 requires-python = ">=3.10"
-dependencies = ["aiohttp>=3.9"]
+dependencies = ["fastapi>=0.100"]
 
 [project.optional-dependencies]
-dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
+dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.24"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
 asyncio_mode = "auto"
+
+[tool.hatch.build.targets.wheel]
+packages = ["."]
